requirements.txt
aiogram==3.13.1
google-generativeai>=0.8.0
Pillow>=10.0.0
